requirements.txt
fastapi
uvicorn
mysql-connector-python
sqlalchemy
pandas
python-dotenv
streamlit
requests
plotly
